chore(vincent-rest-api): Move OpenAPI to sub dir
