+++ packages/vincent-rest-api/src/utils/generate-openapi.ts
@@ -1,10 +1,10 @@
 import fs from 'fs';
 import path from 'path';
-import { registry } from '../lib/baseRegistry';
+import { registry } from '../lib/openApi/baseRegistry';
 import { OpenApiGeneratorV3 } from '@asteasolutions/zod-to-openapi';
-import { addToRegistry as addPolicyToRegistry } from '../lib/policy';
-import { addToRegistry as addToolToRegistry } from '../lib/tool';
-import { addToRegistry as addAppToRegistry } from '../lib/app';
+import { addToRegistry as addPolicyToRegistry } from '../lib/openApi/policy';
+import { addToRegistry as addToolToRegistry } from '../lib/openApi/tool';
+import { addToRegistry as addAppToRegistry } from '../lib/openApi/app';
 
 addAppToRegistry(registry);
 addToolToRegistry(registry);
